+++ CHANGES.rst
@@ -13,6 +13,7 @@ Development:
      ``--bluezenv-progress``.
    - Expose ``bluez_src_dir`` for locating files in the BlueZ source
      tree given by ``--bluez-src-dir``.
+   - Add ``args`` to ``Bluetoothctl`` for extra command-line options.
 
 v0.1.9:
    - Add configurable default VM memory with --vm-mem.